AuthForm to AuthPage
