+++ src/pages/AuthPage/AuthPage.module.css
@@ -1,3 +1,68 @@
+.section {
+  padding-left: 20px;
+  padding-right: 20px;
+  margin-left: auto;
+  margin-right: auto;
+  height: 100vh;
+  background: linear-gradient(#c4c4c400, var(--color-green));
+  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
+  mix-blend-mode: color-burn;
+  position: relative;
+}
+
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+.register_form_window {
+  position: absolute;
+  width: 90vw;
+  padding: 24px;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%) scale(1);
+  background-color: var(--bg-color-form);
+  border-radius: 8px;
+  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
+}
+
+@media screen and (min-width: 375px) {
+  .register_form_window {
+    width: 335px;
+  }
+}
+
+@media screen and (min-width: 768px) {
+  .register_form_window {
+    padding: 40px;
+    width: 424px;
+  }
+}
+
+.form_title {
+  font-family: var(--font-medium);
+  color: #fff;
+  font-size: 18px;
+  line-height: normal;
+  letter-spacing: -0.36px;
+  margin-right: 14px;
+}
+
+.active {
+  font-family: var(--font-medium);
+  color: var(--color-form-light);
+  font-size: 18px;
+  line-height: normal;
+  letter-spacing: -0.36px;
+  margin-right: 14px;
+}
+
 .form {
   display: flex;
   flex-direction: column;
